Rename XOR.ts to Password.ts
