+++ functions/Password.ts
